\name{coerce_ABUSE_DEP_to_nice_string}
\alias{coerce_ABUSE_DEP_to_nice_string}
\title{Converts the processed Abuse and Depression rates into nice strings.}
\usage{
coerce_ABUSE_DEP_to_nice_string(x)
}
\arguments{
  \item{x}{Processed Abuse or Depression Rates.}
}
\description{
Converts the processed Abuse and Depression rates into nice
strings.
}

