\name{coerce_condom_usage_to_nice_string}
\alias{coerce_condom_usage_to_nice_string}
\title{Converts the processed Condom Usage rates into nice strings.}
\usage{
coerce_condom_usage_to_nice_string(x)
}
\arguments{
  \item{x}{Processed Condom Usage Rates.}
}
\description{
Converts the processed Condom Usage rates into nice
strings.
}

