\name{coerce_EXCEL_col_header_to_df}
\alias{coerce_EXCEL_col_header_to_df}
\title{Maps the column headers used in the "child outcomes data.xlsx" file to a data.frame}
\usage{
coerce_EXCEL_col_header_to_df(x)
}
\arguments{
  \item{x}{The column header split on the periods}
}
\description{
Maps the column headers used in the "child outcomes
data.xlsx" file to a data.frame
}

