\docType{data}
\name{DropOutRates}
\alias{DropOutRates}
\title{The dropout probabilities}
\format{One probability for a given age and socio economic status per row}
\usage{
data(DropOutRates)
}
\description{
The data was obtained from the 'child outcomes data.xlsx'
file in the worksheet 'Schooling Data'. Cells B5:B16 and
H5:H16 were used. Note that this was calculated in that
sheet from other source data. See A20:B41
}
\details{
Column names were chosen to match the parameter names used
in the modgen model.
}
\keyword{datasets}

