\name{format_AbusedDepressionRates}
\alias{format_AbusedDepressionRates}
\title{Formats the AbusedDepressionRates into formats the ModGen can use.}
\usage{
format_AbusedDepressionRates(AbusedDepressionRates)
}
\arguments{
  \item{AbuseRates}{The dataset to be mapped and
  formatted.}
}
\description{
Also maps the X0-X5 disease states onto the disease status
used by the ModGen model
}
\details{
Just calls the format_AbuseRates function.
}

