\name{format_CondomUsageRates}
\alias{format_CondomUsageRates}
\title{Formats the CondomUsageRates into formats the ModGen can use. (Abused and Not Abused)}
\usage{
format_CondomUsageRates(CondomUsageRates)
}
\arguments{
  \item{CondomUsageRates}{The dataset to be mapped and
  formatted.}
}
\description{
Also maps the X0-X5 disease states onto the disease status
used by the ModGen model
}

