\name{format_DebutRates}
\alias{format_DebutRates}
\title{Formats the Debut Rates into formats the ModGen can use. (Abused and Not Abused)}
\usage{
format_DebutRates(DebutRates)
}
\arguments{
  \item{DebutRates}{The dataset to be mapped and
  formatted.}
}
\description{
Also maps the X0-X5 disease states onto the disease status
used by the ModGen model
}

