\name{format_DropOutRates}
\alias{format_DropOutRates}
\title{Formats the DropOutRates into formats that modgen can use}
\usage{
format_DropOutRates(DropOutRates)
}
\arguments{
  \item{DropOutRates}{The dataset to be formatted}
}
\description{
Formats the DropOutRates into formats that modgen can use
}

