\name{format_NotAbusedDepressionRates}
\alias{format_NotAbusedDepressionRates}
\title{Formats the NotAbusedDepressionRates into formats the ModGen can use.}
\usage{
format_NotAbusedDepressionRates(NotAbusedDepressionRates)
}
\arguments{
  \item{AbuseRates}{The dataset to be mapped and
  formatted.}
}
\description{
Also maps the X0-X5 disease states onto the disease status
used by the ModGen model
}
\details{
Just calls the format_AbuseRates function.
}

